Adds websocket support to connection handling
This commit add WebSocket support. This required refactoring of Beauty
internal that was done in many commits a bit back and forth.
Also the new WebSocket API had muliple refactoring before settling on
the currect version. So in the end, these commits was finally squashed
as single commit that provides the final WebSocket support in Beauty.