Adds WsEncoder
This commit provides WsEncoder that encapsulates the websocket frame
encoding. The WsEncoder is prepared for both server and client use. As
client use requires masking with a random 32-bit number, also a IRandom
interface with default/fast implementations is provided. The IRandom
interface enables hardware accelerated random generators to be used.

It also moves ownership of the Reply::content_ buffer to the
Connection class. Thus it can be used for WebSocket too and not just
http replies.